Feat: add primitives pallet
