+++ primitives/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+authors = ['InvArchitects <https://github.com/InvArch>']
+description = 'InvArch primitives for InvArchh Pallet Library'
+edition = '2018'
+homepage = 'https://invarch.io'
+license = 'GPLv3'
+name = 'invarch-primitives'
+repository = 'https://github.com/InvArch/InvArch-Pallet-Library/ipt'
+version = '0.1.0-dev'
+
+[dependencies]
+serde = { version = "1.0.124", optional = true }
+codec = { package = "parity-scale-codec", version = "2.2.0", default-features = false }
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.9" }
+frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.9", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.9", default-features = false }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.9", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+    "serde",
+    "codec/std",
+    "sp-core/std",
+    "frame-system/std",
+    "sp-std/std",
+]
